+++ PCL.Core/IO/Net/Http/Client/Request/HttpSenderExtension.cs
@@ -1,0 +1,37 @@
+using PCL.Core.App;
+using PCL.Core.Utils.Exts;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace PCL.Core.IO.Net.Http.Client.Request;
+
+public static class HttpSenderExtension
+{
+    extension (HttpRequestMessage requestMessage)
+    {
+        public async Task<HttpResponseMessage> SendAsync(bool withLauncherMetadata = true, HttpCompletionOption httpCompletionOption = HttpCompletionOption.ResponseContentRead, int retryTimes = 3, CancellationToken cancellationToken = default)
+        {
+            using var resq = requestMessage;
+
+            if (withLauncherMetadata)
+            {
+                requestMessage.Headers.TryAddWithoutValidation("User-Agent", $"PCL-Community/PCL2-CE/{Basics.VersionName} (pclc.cc)");
+                requestMessage.Headers.TryAddWithoutValidation("Referer", $"https://{Basics.VersionCode}.ce.open.pcl2.server/");
+            }
+
+            return await NetworkService.GetRetryPolicy(retryTimes)
+                .ExecuteAsync(async token =>
+                {
+                    using var requestCopy = await resq
+                        .CloneAsync()
+                        .ConfigureAwait(false);
+                    return await NetworkService
+                        .GetClient()
+                        .SendAsync(requestCopy, httpCompletionOption, token)
+                        .ConfigureAwait(false);
+                }, cancellationToken)
+                .ConfigureAwait(false);
+        }
+    }
+}
