+++ PCL.Core/IO/Net/NetworkService.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Net;
 using System.Net.Http;
-using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using PCL.Core.App;
 using PCL.Core.App.IoC;
